Form to add car work nice. Shows errors of duplicate in the db.
chore: Update dependencies and add django_bootstrap5 to installed apps

This commit updates the dependencies in the `requirements.txt` file to include the latest stable versions of `asgiref`, `Django`, and `psycopg2-binary`. Additionally, it adds `django_bootstrap5` to the `INSTALLED_APPS` list in the `app/settings.py` file. This package provides Bootstrap 5 integration for the Django application.