Now views and urls are add, search, edit, delete
chore: Update URLs and views for adding and editing cars

This commit updates the URLs and views in the catalog app to improve the functionality for adding and editing cars. The URLs for adding and editing cars have been changed to '/add/' and '/edit/<int:id>/', respectively. The corresponding views have also been updated to use the new URLs and provide a better user experience.