feat: Add utils folder to .gitignore
This commit updates the .gitignore file to exclude the `utils` folder from version control. The `utils` folder contains JSON, PNG, PDF, and YAML files that are not necessary to track in the repository. Ignoring these files will prevent them from being accidentally committed and improve the cleanliness of the repository.